Split lib/kind.rb into several files
