+++ lib/kind/error.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Kind
+  class Error < TypeError
+    def initialize(klass, object)
+      super("#{object.inspect} expected to be a kind of #{klass}")
+    end
+  end
+end
