+++ lib/kind/undefined.rb
@@ -1,0 +1,27 @@
+# frozen_string_literal: true
+
+module Kind
+  Undefined = Object.new.tap do |undefined|
+    def undefined.inspect
+      @inspect ||= 'Undefined'.freeze
+    end
+
+    def undefined.to_s
+      inspect
+    end
+
+    def undefined.clone
+      self
+    end
+
+    def undefined.dup
+      clone
+    end
+
+    def undefined.default(value, default)
+      return self if value != self
+
+      default.respond_to?(:call) ? default.call : default
+    end
+  end
+end
